requirements.txt
streamlit==1.12.0
opencv-python-headless==4.5.3.56
numpy==1.22.4
Pillow==9.1.1